Refactor #28: View를 직접적으로 그리지 않는 Extension을 Utility로 이동
- navigationBarBackButton extension은 View를 직접 그리기 때문에 Buttons 폴더로 이동
- Utility 의존성 연결 (DesignSystem, MarkPlace, Authorization)
