+++ Umat/Projects/Feature/MarkPlace/Project.swift
@@ -20,6 +20,7 @@ let project = Project.makeAppModule(
     dependencies: [
         .shared(implements: .designSystem),
         .shared(implements: .entity),
+        .shared(implements: .utility)
     ],
     sources: ["Sources/**"],
     infoPlist: .file(path: "../../Support/Info.plist")
