+++ Umat/Projects/Shared/DesignSystem/Project.swift
@@ -11,7 +11,7 @@ let project = Project.makeAppModule(
     product: .staticFramework,
     settings:  .settings(),
     dependencies: [
-        
+        .shared(implements: .utility)
     ],
     sources: ["Sources/**"],
     resources: ["Resources/**"],
