+++ Umat/Projects/Shared/DesignSystem/Sources/Component/Buttons/NavigationBarBackButton.swift
@@ -2,7 +2,7 @@
 //  NavigationBarBackButton.swift
 //  DesignSystem
 //
-//  Created by 지준용 on 2/12/24.
+//  Created by 지준용 on 2/8/24.
 //  Copyright © 2024 KYUNG MIN CHOI. All rights reserved.
 //
 
@@ -20,5 +20,19 @@ public struct NavigationBarBackButton: View {
             Image.icons(.ic_arrow_back_filled)
                 .resizable()
         }
+    }
+}
+
+
+public extension View {
+    @ViewBuilder
+    func navigationBarBackButton() -> some View {
+        self
+            .navigationBarBackButtonHidden()
+            .toolbar {
+                ToolbarItem(placement: .topBarLeading) {
+                    NavigationBarBackButton()
+                }
+            }
     }
 }
