+++ Umat/Projects/Shared/Utility/Sources/Extension/SwiftUI+Extension/View+Extension/View+shape.swift
@@ -1,0 +1,28 @@
+//
+//  View+shape.swift
+//  Utility
+//
+//  Created by 지준용 on 2/24/24.
+//
+
+import SwiftUI
+
+
+struct RoundedCorner: Shape {
+    var radius: CGFloat = .infinity
+    var corners: UIRectCorner = .allCorners
+
+    func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect,
+                                byRoundingCorners: corners,
+                                cornerRadii: CGSize(width: radius,
+                                                    height: radius))
+        return Path(path.cgPath)
+    }
+}
+
+public extension View {
+    func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
+        clipShape( RoundedCorner(radius: radius, corners: corners) )
+    }
+}
